+++ test/test_web.rb
@@ -164,7 +164,7 @@ describe Sidekiq::Web do
 
     Sidekiq.redis do |conn|
       refute conn.smembers('queues').include?('foo')
-      refute conn.exists('queue:foo')
+      refute conn.exists?('queue:foo')
     end
   end