+++ Changes.md
@@ -5,6 +5,7 @@
 HEAD
 ---------
 
+- Fix: Do not connect to redis at ruby vm exit when not needed. [#4502] 
 - Remove Redis connection naming [#4479]
 
 6.0.6