+++ Changes.md
@@ -5,8 +5,8 @@
 HEAD
 ---------
 
-- Fix: Do not connect to redis at ruby vm exit when not needed. [#4502] 
 - Refactor systemd integration to work better with custom binaries [#4511]
+- Don't connect to Redis at process exit if not needed [#4502]
 - Remove Redis connection naming [#4479]
 
 6.0.6