+++ Changes.md
@@ -5,6 +5,8 @@
 Unreleased
 ---------
 
+- Web UI - Dark Mode fixes [#4543, natematykiewicz]
+- Ensure `Rack::ContentLength` is loaded as middleware for correct Web UI responses [#4541]
 - Avoid exception dumping SSL store in Redis connection logging [#4532]
 
 6.0.7